requirements.txt
qrcode
flask 
pillow